fix sorting icons, translations
