+++ alv-portal-ui/src/app/competence-catalog/shared/sort/competence-catalog-sort-modal/competence-catalog-sort-modal.component.ts
@@ -1,8 +1,9 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { of } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { CompetenceCatalogSortValue, SortIcon, SortType } from '../../constants';
+import { SelectableOption } from '../../../../shared/forms/input/selectable-option.model';
+import { CompetenceCatalogSortValue, SortIcon, SortType } from '../../shared-competence-catalog.types';
 
 @Component({
   selector: 'alv-competence-catalog-sort-modal',
@@ -16,24 +17,12 @@ export class CompetenceCatalogSortModalComponent implements OnInit {
 
   form: FormGroup;
 
-  sortTypes$ = of([
-    {
-      label: 'portal.competence-catalog.sorting.created-date-desc',
-      value: SortType.CREATED_DATE_DESC
-    },
-    {
-      label: 'portal.competence-catalog.sorting.created-date-asc',
-      value: SortType.CREATED_DATE_ASC
-    },
-    {
-      label: 'portal.competence-catalog.sorting.alpha-desc',
-      value: SortType.ALPHA_DESC
-    },
-    {
-      label: 'portal.competence-catalog.sorting.alpha-asc',
-      value: SortType.ALPHA_ASC
-    }
-  ]);
+  sortTypes$: Observable<SelectableOption[]> = of(Object.values(SortType)
+    .map(s => ({
+      label: 'portal.competence-catalog.sorting.' + s,
+      value: s
+    }))
+  );
 
   constructor(private fb: FormBuilder,
               public activeModal: NgbActiveModal) {
@@ -43,6 +32,7 @@ export class CompetenceCatalogSortModalComponent implements OnInit {
     this.form = this.fb.group({
       sort: [this.currentSorting.type, Validators.required]
     });
+
   }
 
   dismiss() {
@@ -54,16 +44,16 @@ export class CompetenceCatalogSortModalComponent implements OnInit {
     let icon: SortIcon;
     switch (type) {
       case SortType.CREATED_DATE_DESC:
-        icon = SortIcon.NUMERIC_DESC;
-        break;
-      case SortType.CREATED_DATE_ASC:
         icon = SortIcon.NUMERIC_UP;
         break;
+      case SortType.CREATED_DATE_ASC:
+        icon = SortIcon.NUMERIC_DOWN;
+        break;
       case SortType.ALPHA_DESC:
-        icon = SortIcon.ALPHA_DESC;
+        icon = SortIcon.ALPHA_UP;
         break;
       case SortType.ALPHA_ASC:
-        icon = SortIcon.ALPHA_ASC;
+        icon = SortIcon.ALPHA_DOWN;
         break;
     }
     const updatedSorting: CompetenceCatalogSortValue = { type, icon };
