+++ alv-portal-ui/translations.csv
@@ -1162,10 +1162,10 @@ portal.application-documents.document-type.CV,Lebenslauf,Lebenslauf,Curriculum v
 portal.application-documents.document-type.LETTER_OF_APPLICATION,Bewerbungsschreiben,Bewerbungsschreiben,Lettre de motivation,Domande d'impiego,
 portal.application-documents.document-type.TESTIMONIAL,Arbeitszeugnisse,Arbeitszeugnisse,Certificats de travail,Certificati di lavoro,
 portal.application-documents.document-type.CERTIFICATES,Diplome / Zertifikate,Diplome / Zertifikate,Diplômes/certificats,Diplomi / Certificati,
-portal.competence-catalog.sorting.created-date-desc,Erstellungsdatum absteigend,Creation date descending,Date de création décroissante,Data di creazione decrescente,
-portal.competence-catalog.sorting.created-date-asc,Erstellungsdatum aufsteigend,Creation date ascending,Date de création croissant,Data di creazione crescente,
-portal.competence-catalog.sorting.alpha-desc,Alphabetisch absteigend,Alphabetically descending,Ordre alphabétique descendant,In ordine alfabetico decrescente,
-portal.competence-catalog.sorting.alpha-asc,Alphabetisch aufsteigend,Alphabetically ascending,Ordre alphabétique croissant,In ordine alfabetico ascendente,
+portal.competence-catalog.sorting.CREATED_DATE_DESC,Erstellungsdatum absteigend,Creation date descending,Date de création décroissante,Data di creazione decrescente,
+portal.competence-catalog.sorting.CREATED_DATE_ASC,Erstellungsdatum aufsteigend,Creation date ascending,Date de création croissant,Data di creazione crescente,
+portal.competence-catalog.sorting.ALPHA_DESC,Alphabetisch absteigend,Alphabetically descending,Ordre alphabétique descendant,In ordine alfabetico decrescente,
+portal.competence-catalog.sorting.ALPHA_ASC,Alphabetisch aufsteigend,Alphabetically ascending,Ordre alphabétique croissant,In ordine alfabetico ascendente,
 portal.competence-catalog.competence-elements.browser-title,Kompetenz-Bestandteile,Skill components,Composants de compétence,Elementi della competenza,
 portal.competence-catalog.competence-sets.browser-title,Kompetenz-Sets,Skill sets,Ensembles de compétences,Set di competenze,
 portal.competence-catalog.element-type.KNOW_HOW,Handlungswissen,Know-how,Savoirs-faire,Competenze operative,
